+++ config-builder/src/main/java/com/ca/apim/gateway/cagatewayconfig/bundle/builder/AuthenticationAssertionBuilder.java
@@ -20,7 +20,7 @@ public class AuthenticationAssertionBuilder implements PolicyAssertionBuilder {
             final String idProviderName =
                     idProviderNameElement.getAttributes().getNamedItem(STRING_VALUE).getTextContent();
             if(!idProviderName.equals(INTERNAL_IDP_NAME)){
-                idProviderNameElement.setAttribute(STRING_VALUE, bundle.applyUniqueName(idProviderName, ENVIRONMENT, false));
+                idProviderNameElement.setAttribute(STRING_VALUE, bundle.applyUniqueName(idProviderName, ENVIRONMENT));
             }
         }
     }
