+++ config-builder/src/main/java/com/ca/apim/gateway/cagatewayconfig/bundle/builder/CassandraAssertionBuilder.java
@@ -16,7 +16,7 @@ public class CassandraAssertionBuilder implements PolicyAssertionBuilder {
         if (connectionNameElement != null) {
             final String connectionName =
                     connectionNameElement.getAttributes().getNamedItem(STRING_VALUE).getTextContent();
-            connectionNameElement.setAttribute(STRING_VALUE, bundle.applyUniqueName(connectionName, ENVIRONMENT, false));
+            connectionNameElement.setAttribute(STRING_VALUE, bundle.applyUniqueName(connectionName, ENVIRONMENT));
         }
     }
