+++ config-builder/src/main/java/com/ca/apim/gateway/cagatewayconfig/bundle/builder/GenericEntityBuilder.java
@@ -41,12 +41,12 @@ public class GenericEntityBuilder implements EntityBuilder {
         switch (bundleType) {
             case DEPLOYMENT:
                 return genericEntities.entrySet().stream()
-                        .map(entry -> getEntityWithOnlyMapping(EntityTypes.GENERIC_TYPE, bundle.applyUniqueName(entry.getKey(), BundleType.ENVIRONMENT, false),
+                        .map(entry -> getEntityWithOnlyMapping(EntityTypes.GENERIC_TYPE, bundle.applyUniqueName(entry.getKey(), BundleType.ENVIRONMENT),
                                 generateId(entry.getValue())))
                         .collect(Collectors.toList());
             case ENVIRONMENT:
                 return genericEntities.entrySet().stream()
-                        .map(e -> buildGenericEntity(bundle.applyUniqueName(e.getKey(), bundleType, false), e.getValue(), document))
+                        .map(e -> buildGenericEntity(bundle.applyUniqueName(e.getKey(), bundleType), e.getValue(), document))
                         .collect(Collectors.toList());
             default:
                 throw new EntityBuilderException("Unknown bundle type: " + bundleType);
