+++ config-builder/src/main/java/com/ca/apim/gateway/cagatewayconfig/bundle/builder/JmsAssertionBuilder.java
@@ -16,7 +16,7 @@ public class JmsAssertionBuilder implements PolicyAssertionBuilder {
         if (connectionNameElement != null) {
             final String connectionName =
                     connectionNameElement.getAttributes().getNamedItem(STRING_VALUE).getTextContent();
-            connectionNameElement.setAttribute(STRING_VALUE, bundle.applyUniqueName(connectionName, ENVIRONMENT, false));
+            connectionNameElement.setAttribute(STRING_VALUE, bundle.applyUniqueName(connectionName, ENVIRONMENT));
         }
     }
