+++ config-builder/src/test/java/com/ca/apim/gateway/cagatewayconfig/bundle/builder/BundleEntityBuilderTest.java
@@ -15,6 +15,7 @@ import com.ca.apim.gateway.cagatewayconfig.util.entity.EntityTypes;
 import com.ca.apim.gateway.cagatewayconfig.util.gateway.MappingActions;
 import com.ca.apim.gateway.cagatewayconfig.util.xml.DocumentTools;
 import org.jetbrains.annotations.NotNull;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.testcontainers.shaded.com.google.common.collect.ImmutableMap;
 import org.w3c.dom.Document;
@@ -53,6 +54,25 @@ class BundleEntityBuilderTest {
 
     private static final ProjectInfo projectInfo = new ProjectInfo("my-bundle", "my-bundle-group", "1.0");
 
+    private PolicyXMLBuilder policyXMLBuilder;
+
+    @BeforeEach
+    void beforeEach() {
+        Set<PolicyAssertionBuilder> policyAssertionBuilders = new HashSet<>();
+        Reflections reflections = new Reflections();
+        reflections.getSubTypesOf(PolicyAssertionBuilder.class).forEach(e -> {
+            try {
+                policyAssertionBuilders.add(e.newInstance());
+            } catch (InstantiationException ex) {
+                ex.printStackTrace();
+            } catch (IllegalAccessException ex) {
+                ex.printStackTrace();
+            }
+        });
+        policyXMLBuilder = new PolicyXMLBuilder(policyAssertionBuilders);
+
+    }
+
     // This class is covered by testing others, so a simple testing is enough here.
     @Test
     void build() {
@@ -79,7 +99,7 @@ class BundleEntityBuilderTest {
     @Test
     void testEncassAnnotatedBundle() {
         FolderEntityBuilder folderBuilder = new FolderEntityBuilder(ID_GENERATOR);
-        PolicyEntityBuilder policyBuilder = new PolicyEntityBuilder(DocumentTools.INSTANCE, ID_GENERATOR);
+        PolicyEntityBuilder policyBuilder = new PolicyEntityBuilder(DocumentTools.INSTANCE, ID_GENERATOR, policyXMLBuilder);
         EncassEntityBuilder encassBuilder = new EncassEntityBuilder(ID_GENERATOR);
 
         Bundle bundle = new Bundle();
@@ -132,7 +152,7 @@ class BundleEntityBuilderTest {
     @Test
     void testEncassAnnotatedRedeployableBundle() {
         FolderEntityBuilder folderBuilder = new FolderEntityBuilder(ID_GENERATOR);
-        PolicyEntityBuilder policyBuilder = new PolicyEntityBuilder(DocumentTools.INSTANCE, ID_GENERATOR);
+        PolicyEntityBuilder policyBuilder = new PolicyEntityBuilder(DocumentTools.INSTANCE, ID_GENERATOR, policyXMLBuilder);
         EncassEntityBuilder encassBuilder = new EncassEntityBuilder(ID_GENERATOR);
 
         Bundle bundle = new Bundle();
@@ -257,7 +277,7 @@ class BundleEntityBuilderTest {
     public void testAnnotatedEncassDeleteBundle() {
         BundleEntityBuilder builder = createBundleEntityBuilder();
 
-        Bundle bundle = createBundle(ENCASS_POLICY_WITH_ENV_DEPENDENCIES, true, true, false);
+        Bundle bundle = createBundle(ENCASS_POLICY_WITH_ENV_DEPENDENCIES, true, true, false, projectInfo);
         Encass encass = buildTestEncassWithAnnotation(TEST_GUID, TEST_ENCASS_POLICY, false);
         bundle.putAllEncasses(ImmutableMap.of(TEST_ENCASS, encass));
 
@@ -286,7 +306,7 @@ class BundleEntityBuilderTest {
         assertEquals(EntityTypes.POLICY_TYPE, getSingleChildElementTextContent(item1, TYPE));
         assertNotNull(getSingleChildElement(item1, RESOURCE));
         final Element item2 = itemList.get(0);
-        assertEquals(TEST_ENCASS, getSingleChildElementTextContent(item2, NAME));
+        assertEquals("::" + projectInfo.getGroupName() + "." + TEST_ENCASS_ANNOTATION_NAME + "::" + TEST_ENCASS + "::" + projectInfo.getVersion(), getSingleChildElementTextContent(item2, NAME));
         assertEquals(EntityTypes.ENCAPSULATED_ASSERTION_TYPE, getSingleChildElementTextContent(item2, TYPE));
         assertNotNull(getSingleChildElement(item2, RESOURCE));
 
@@ -308,7 +328,7 @@ class BundleEntityBuilderTest {
     public void testAnnotatedEncassDeleteBundle_WithPolicyFragment() {
         BundleEntityBuilder builder = createBundleEntityBuilder();
 
-        Bundle bundle = createBundleWithPolicyFragment(false);
+        Bundle bundle = createBundleWithPolicyFragment(false, projectInfo);
         Encass encass = buildTestEncassWithAnnotation(TEST_GUID, TEST_ENCASS_POLICY, false);
         bundle.putAllEncasses(ImmutableMap.of(TEST_ENCASS, encass));
 
@@ -342,7 +362,7 @@ class BundleEntityBuilderTest {
         assertEquals(EntityTypes.POLICY_TYPE, getSingleChildElementTextContent(item2, TYPE));
         assertNotNull(getSingleChildElement(item2, RESOURCE));
         final Element item3 = itemList.get(0);
-        assertEquals(TEST_ENCASS, getSingleChildElementTextContent(item3, NAME));
+        assertEquals("::" + projectInfo.getGroupName() + "." + TEST_ENCASS_ANNOTATION_NAME + "::" + TEST_ENCASS + "::" + projectInfo.getVersion(), getSingleChildElementTextContent(item3, NAME));
         assertEquals(EntityTypes.ENCAPSULATED_ASSERTION_TYPE, getSingleChildElementTextContent(item3, TYPE));
         assertNotNull(getSingleChildElement(item3, RESOURCE));
 
@@ -369,7 +389,7 @@ class BundleEntityBuilderTest {
         final Element mapping3 = mappingItemList.get(0);
         final Element mapping3Properties = getSingleChildElement(mapping3, PROPERTIES);
         propertyValues = getChildElementsTextContents(mapping3Properties, PROPERTY);
-        assertTrue(propertyValues.contains(TEST_ENCASS));
+        assertTrue(propertyValues.contains("::" + projectInfo.getGroupName() + "." + TEST_ENCASS_ANNOTATION_NAME + "::" + TEST_ENCASS + "::" + projectInfo.getVersion()));
         assertEquals(MappingActions.DELETE, mapping3.getAttribute("action"));
         assertEquals(EntityTypes.ENCAPSULATED_ASSERTION_TYPE, mapping3.getAttribute("type"));
     }
@@ -378,7 +398,7 @@ class BundleEntityBuilderTest {
     public void testAnnotatedEncassDeleteBundle_WithReusablePolicyFragment_RedeployableBundle() {
         BundleEntityBuilder builder = createBundleEntityBuilder();
 
-        Bundle bundle = createBundleWithPolicyFragment(false);
+        Bundle bundle = createBundleWithPolicyFragment(false, projectInfo);
         Encass encass = buildTestEncassWithAnnotation(TEST_GUID, TEST_ENCASS_POLICY, true);
         bundle.putAllEncasses(ImmutableMap.of(TEST_ENCASS, encass));
 
@@ -412,7 +432,7 @@ class BundleEntityBuilderTest {
         assertEquals(EntityTypes.POLICY_TYPE, getSingleChildElementTextContent(item2, TYPE));
         assertNotNull(getSingleChildElement(item2, RESOURCE));
         final Element item3 = itemList.get(0);
-        assertEquals(TEST_ENCASS, getSingleChildElementTextContent(item3, NAME));
+        assertEquals("::" + projectInfo.getGroupName() + "." + TEST_ENCASS_ANNOTATION_NAME + "::" + TEST_ENCASS + "::" + projectInfo.getVersion(), getSingleChildElementTextContent(item3, NAME));
         assertEquals(EntityTypes.ENCAPSULATED_ASSERTION_TYPE, getSingleChildElementTextContent(item3, TYPE));
         assertNotNull(getSingleChildElement(item3, RESOURCE));
 
@@ -439,7 +459,7 @@ class BundleEntityBuilderTest {
         final Element mapping3 = mappingItemList.get(0);
         final Element mapping3Properties = getSingleChildElement(mapping3, PROPERTIES);
         propertyValues = getChildElementsTextContents(mapping3Properties, PROPERTY);
-        assertTrue(propertyValues.contains(TEST_ENCASS));
+        assertTrue(propertyValues.contains("::" + projectInfo.getGroupName() + "." + TEST_ENCASS_ANNOTATION_NAME + "::" + TEST_ENCASS + "::" + projectInfo.getVersion()));
         assertEquals(MappingActions.DELETE, mapping3.getAttribute("action"));
         assertEquals(EntityTypes.ENCAPSULATED_ASSERTION_TYPE, mapping3.getAttribute("type"));
     }
@@ -448,7 +468,7 @@ class BundleEntityBuilderTest {
     public void testAnnotatedEncassDeleteBundle_WithReusablePolicyFragment() {
         BundleEntityBuilder builder = createBundleEntityBuilder();
 
-        Bundle bundle = createBundleWithPolicyFragment(true);
+        Bundle bundle = createBundleWithPolicyFragment(true, projectInfo);
         Encass encass = buildTestEncassWithAnnotation(TEST_GUID, TEST_ENCASS_POLICY, false);
         bundle.putAllEncasses(ImmutableMap.of(TEST_ENCASS, encass));
 
@@ -478,7 +498,7 @@ class BundleEntityBuilderTest {
         assertEquals(EntityTypes.POLICY_TYPE, getSingleChildElementTextContent(item2, TYPE));
         assertNotNull(getSingleChildElement(item2, RESOURCE));
         final Element item3 = itemList.get(0);
-        assertEquals(TEST_ENCASS, getSingleChildElementTextContent(item3, NAME));
+        assertEquals("::" + projectInfo.getGroupName() + "." + TEST_ENCASS_ANNOTATION_NAME + "::" + TEST_ENCASS + "::" + projectInfo.getVersion(), getSingleChildElementTextContent(item3, NAME));
         assertEquals(EntityTypes.ENCAPSULATED_ASSERTION_TYPE, getSingleChildElementTextContent(item3, TYPE));
         assertNotNull(getSingleChildElement(item3, RESOURCE));
 
@@ -500,7 +520,7 @@ class BundleEntityBuilderTest {
         final Element mapping3 = mappingItemList.get(0);
         final Element mapping3Properties = getSingleChildElement(mapping3, PROPERTIES);
         propertyValues = getChildElementsTextContents(mapping3Properties, PROPERTY);
-        assertTrue(propertyValues.contains(TEST_ENCASS));
+        assertTrue(propertyValues.contains("::" + projectInfo.getGroupName() + "." + TEST_ENCASS_ANNOTATION_NAME + "::" + TEST_ENCASS + "::" + projectInfo.getVersion()));
         assertEquals(MappingActions.DELETE, mapping3.getAttribute("action"));
         assertEquals(EntityTypes.ENCAPSULATED_ASSERTION_TYPE, mapping3.getAttribute("type"));
     }
