+++ environment-creator-application/src/test/java/com/ca/apim/gateway/cagatewayconfig/EnvironmentCreatorApplicationTest.java
@@ -292,7 +292,7 @@ class EnvironmentCreatorApplicationTest {
         Bundle environmentBundle = bundleLoader.load(environmentBundleFile, BundleLoadingOperation.EXPORT);
 
         assertEquals(1, environmentBundle.getIdentityProviders().size());
-        assertNotNull(environmentBundle.getIdentityProviders().get("simple ldap"));
+        assertNotNull(environmentBundle.getIdentityProviders().get("::::simple ldap"));
 
         assertEquals(3, environmentBundle.getListenPorts().size());
         assertNotNull(environmentBundle.getListenPorts().get("Custom HTTPS Port"));
@@ -300,7 +300,7 @@ class EnvironmentCreatorApplicationTest {
         assertNotNull(environmentBundle.getListenPorts().get(DEFAULT_HTTPS_8443));
 
         assertEquals(1, environmentBundle.getJdbcConnections().size());
-        assertNotNull(environmentBundle.getJdbcConnections().get("my-jdbc"));
+        assertNotNull(environmentBundle.getJdbcConnections().get("::::my-jdbc"));
 
         assertEquals(1, environmentBundle.getTrustedCerts().size());
         assertNotNull(environmentBundle.getTrustedCerts().get("my-cert"));
